airflow/requirements.txt
apache-airflow-providers-google
dbt-bigquery
pandas
requests
google-cloud-bigquery
